utils/requirements.txt
mysql-connector==2.1.4
PyYAML==3.12